Add Alamofire framework via Carthage and test it
This uses a local git repo for the Alamofire networking framework; I did it this way so that I can learn how to use third-party frameworks in my projects without requiring a network connection (for example, when working on a project on my laptop in the park). Frameworks can be updated independently at any time I have a connection to the internet, so this is a viable solution.